EKF: move all estimator specific types to common header
