+++ EKF/estimator_base.h
@@ -43,118 +43,8 @@
 #include <matrix/matrix/math.hpp>
 #include <lib/geo/geo.h>
 #include "RingBuffer.h"
-namespace estimator {
-	struct gps_message {
-		uint64_t time_usec;
-	    int32_t lat;                // Latitude in 1E-7 degrees
-	    int32_t lon;                // Longitude in 1E-7 degrees
-	    int32_t alt;                // Altitude in 1E-3 meters (millimeters) above MSL
-	    uint8_t fix_type;           // 0-1: no fix, 2: 2D fix, 3: 3D fix, 4: RTCM code differential, 5: Real-Time
-	    float eph;                  // GPS horizontal position accuracy in m
-	    float epv;                  // GPS vertical position accuracy in m
-	    float sacc;                 // GPS speed accuracy in m/s
-	    uint64_t time_usec_vel; 	// Timestamp for velocity informations
-	    float vel_m_s;              // GPS ground speed (m/s)
-	    float vel_ned[3];           // GPS ground speed NED
-	    bool vel_ned_valid;         // GPS ground speed is valid
-	    uint8_t nsats;              // number of satellites used
-	    float gdop;                 // geometric dilution of precision
-	};
 
-	typedef matrix::Vector<float, 2> Vector2f;
-	typedef matrix::Vector<float, 3> Vector3f;
-	typedef matrix::Quaternion<float> Quaternion;
-	typedef matrix::Matrix<float, 3, 3> Matrix3f;
-
-	struct outputSample {
-		Quaternion  quat_nominal;
-		Vector3f    vel;
-		Vector3f    pos;
-		uint64_t 	time_us;
-	};
-
-	struct imuSample {
-		Vector3f    delta_ang;
-		Vector3f    delta_vel;
-		float       delta_ang_dt;
-		float       delta_vel_dt;
-		uint64_t    time_us;
-	};
-
-	struct gpsSample {
-		Vector2f    pos;
-		float       hgt;
-		Vector3f    vel;
-		uint64_t    time_us;
-	};
-
-	struct magSample {
-		Vector3f    mag;
-		uint64_t    time_us;
-	};
-
-	struct baroSample {
-		float       hgt;
-		uint64_t    time_us;
-	};
-
-	struct rangeSample {
-		float       rng;
-		uint64_t    time_us;
-	};
-
-	struct airspeedSample {
-		float       airspeed;
-		uint64_t    time_us;
-	};
-
-	struct flowSample {
-		Vector2f    flowRadXY;
-		Vector2f    flowRadXYcomp;
-		uint64_t    time_us;
-	};
-
-	struct parameters {
-		float mag_delay_ms = 0.0f;
-		float baro_delay_ms = 0.0f;
-		float gps_delay_ms = 200.0f;
-		float airspeed_delay_ms = 200.0f;
-
-		// input noise
-		float gyro_noise = 0.001f;
-		float accel_noise = 0.1f;
-
-		// process noise
-		float gyro_bias_p_noise = 1e-5f;
-		float accel_bias_p_noise = 1e-3f;
-		float gyro_scale_p_noise = 1e-4f;
-		float mag_p_noise = 1e-2f;
-		float wind_vel_p_noise = 0.05f;
-
-		float gps_vel_noise = 0.05f;
-		float gps_pos_noise = 1.0f;
-		float baro_noise = 0.1f;
-        float baro_innov_gate = 5.0f;       // barometric height innovation consistency gate size in standard deviations
-        float posNE_innov_gate = 5.0f;      // GPS horizontal position innovation consistency gate size in standard deviations
-        float vel_innov_gate = 3.0f;        // GPS velocity innovation consistency gate size in standard deviations
-
-		float mag_heading_noise = 3e-2f;	// measurement noise used for simple heading fusion
-		float mag_declination_deg = 0.0f;	// magnetic declination in degrees
-	    float heading_innov_gate = 3.0f;	// heading fusion innovation consistency gate size in standard deviations
-	    float mag_innov_gate = 3.0f;        // magnetometer fusion innovation consistency gate size in standard deviations
-
-	    // these parameters control the strictness of GPS quality checks used to determine uf the GPS is
-	    // good enough to set a local origin and commence aiding
-	    int gps_check_mask = 21;    // bitmask used to control which GPS quality checks are used
-	    float req_hacc = 5.0f;      // maximum acceptable horizontal position error
-	    float req_vacc = 8.0f;      // maximum acceptable vertical position error
-	    float req_sacc = 1.0f;      // maximum acceptable speed error
-	    int req_nsats = 6;          // minimum acceptable satellite count
-	    float req_gdop = 2.0f;      // maximum acceptable geometric dilution of precision
-	    float req_hdrift = 0.3f;    // maximum acceptable horizontal drift speed
-	    float req_vdrift = 0.5f;    // maximum acceptable vertical drift speed
-	};
-}
+#include "common.h"
 
 using namespace estimator;
 class EstimatorBase
@@ -277,19 +167,6 @@ public:
 
 protected:
 
-	struct stateSample {
-		Vector3f    ang_error;
-		Vector3f    vel;
-		Vector3f    pos;
-		Vector3f    gyro_bias;
-		Vector3f    gyro_scale;
-		float       accel_z_bias;
-		Vector3f    mag_I;
-		Vector3f    mag_B;
-		Vector2f    wind_vel;
-		Quaternion  quat_nominal;
-	} _state;
-
 	parameters _params;		// filter parameters
 
 	static const uint8_t OBS_BUFFER_LENGTH = 10;
@@ -346,16 +223,7 @@ protected:
 	uint64_t _time_last_range;
 	uint64_t _time_last_airspeed;
 
-	// flags capturing information about severe nummerical problems for various fusions
-	struct {
-		bool bad_mag_x: 1;
-		bool bad_mag_y: 1;
-		bool bad_mag_z: 1;
-		bool bad_airspeed: 1;
-		bool bad_sideslip: 1;
-	} _fault_status;
 
-    // initialise variables to default startup values and set time stamps to specified value
-	void initialiseVariables(uint64_t timestamp);
-
+	fault_status_t _fault_status;
+	bool initialise_interface(uint64_t timestamp);
 };
